python: move to Python 3
After many years of hobbling along on Python 2, I was finally forced
to update to Python 3. Archlinux finally bit the bullet and dropped
all Python 2 packages, which made installing pywingo enough of a
headache to cause me to just port the damn thing.

As usual, the tricky parts were in dealing with differences in strings.